requirements.txt
aiofiles==0.7.0
aiosqlite==0.16.1
asgiref==3.3.4
asyncpg==0.23.0
atlastk==0.12.10
click==8.0.1
ecdsa==0.17.0
envparse==0.2.0
fastapi==0.65.2
h11==0.12.0
iso8601==0.1.14
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.8.2
pypika-tortoise==0.1.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
rsa==4.7.2
six==1.16.0
starlette==0.14.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
uvicorn==0.14.0
bcrypt==3.2.0